Add time zone offset tags, fix --skip-existing
Brightwheel sends times in UTC, but Exif tags originally didn't include
any time zone info, and most apps interpret the tags as local time.
This leads to Google Photos putting the photos in the wrong place in
my stream.  With this change, we add some newer Exif tags to specify
that the existing tags should be interpreted as UTC.  (It would also be
possible to actually convert to the system's local time, but that would
require depending on Python 3.6 or the tzinfo library, and the files
we create would not be identical across systems or repeated runs.)

Also, fix minor problems that prevented --skip-existing from working.