Download additional images and raw API responses (#30)
Just had a child graduate from a Brightwheel school, and this script was
great for getting all the photos. Thank you for making it!

I added a couple of minor additions that could be useful for others:
1. I discovered that some images are attached to activity types like
`ac_incident`. I wanted to get every possible image, so I changed the
activity generator to get all activities, not just the image and video
activities.
2. Some day I want to have make a little timeline of the all the
Brightwheel activities, so I also changed the script to dump the raw
responses as newline-delimited JSON objects.

---------

Co-authored-by: Johnathan Gilday <me@johnathangilday.com>